requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.0.4
djangorestframework==3.14.0
gunicorn==20.1.0
joblib==1.2.0
numpy==1.23.5
pandas==1.5.2
psycopg2==2.8.6
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
tomli==2.0.1
whitenoise==6.2.0
